// src/types/doctor.types.ts
enum Gender {
    MALE = "MALE",
    FEMALE = "FEMALE",
    OTHER = "OTHER"
}

export enum UserStatus {
    ACTIVE = "ACTIVE",
  BLOCKED = "BLOCKED",
  DELETED = "DELETED",
}


export interface IDoctor {
    id: number;
    name: string;
    email: string;
    profilePhoto?: string;
    contactNumber?: string;
    address?: string;
    registrationNumber: string;
    experience?: number;
    gender: Gender;
    appointmentFee: number;
    qualification: string;
    currentWorkingPlace: string;
    designation: string;
    averageRating: number;
    createdAt: Date;
    user : {
        status : UserStatus
    };
    specialties : Array<{
        specialtyId : string;
        doctorId : string;
        specialty : {
            id : string;
            title : string;
            icon : string;
        }
    }>
}